Keep only tests related to ERC in behaviour
